rename validator to assumptions
