+++ tests/model_utils/test_assumpts.py
@@ -13,14 +13,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Testing the metadata validator."""
+"""Testing the assumptions module."""
 
 from contextlib import nullcontext
 from pathlib import Path
 
 import pytest
 
-from metldata.model_utils.model_validator import (
+from metldata.model_utils.assumptions import (
     MetadataModelAssumptionError,
     check_metadata_model_assumption,
 )
